Deserialization improvement
The Deserialization method did not set the data into the actual game element parameters. It is fixed in this commit.

+++ BrackeysGamejamFinal/Assets/Scripts/Data Management/Serialization/Game Elements Data/EnemyData.cs
@@ -17,7 +17,7 @@ public class EnemyData
     public float armor;
     public float maxArmor;
     public WeaknessType weakness;
-    public float weaknessFactor;
+    public int weaknessFactor;
     public int fireAttack;
     public int waterAttack;
     public int windAttack;
